backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic==2.11.7
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
jmcomic==2.6.4
Pillow==11.3.0
PyYAML==6.0.2
img2pdf==0.6.1
apscheduler==3.11.0
